+++ pyqvd/__init__.py
@@ -3,4 +3,4 @@ Root package of the PyQVD library. Contains the main classes and functions to re
 """
 
 from .qvd import (QvdTable, QvdValue, IntegerValue, DoubleValue, StringValue, DualIntegerValue,
-                  DualDoubleValue, TimeValue, DateValue, TimestampValue, IntervalValue)
+                  DualDoubleValue, TimeValue, DateValue, TimestampValue, IntervalValue, MoneyValue)