+++ tests/io/writer_test.py
@@ -4,6 +4,7 @@ Tests the functionality related to persisting files.
 
 import os
 import datetime as dt
+from decimal import Decimal
 from pyqvd import QvdTable
 
 def test_write_qvd_file(tmp_path):
@@ -12,13 +13,13 @@ def test_write_qvd_file(tmp_path):
     written to file successfully.
     """
     raw_df = {
-        "columns": ["Key", "Value", "Timestamp", "Duration"],
+        "columns": ["Key", "Value", "Timestamp", "Duration", "Salary"],
         "data": [
-            [1, "A", dt.datetime(2021, 1, 1, 0, 0, 0), dt.timedelta(days=1, seconds=3620)],
-            [2, "B", dt.datetime(2021, 1, 2, 0, 0, 0), dt.timedelta(days=2, seconds=7240)],
-            [3, "C", dt.datetime(2021, 1, 3, 0, 0, 0), dt.timedelta(days=3, seconds=7200)],
-            [4, "D", dt.datetime(2021, 1, 4, 0, 0, 0), dt.timedelta(days=4, seconds=10800)],
-            [5, "E", dt.datetime(2021, 1, 5, 0, 0, 0), dt.timedelta(days=5, seconds=14400)],
+            [1, "A", dt.datetime(2021, 1, 1, 0, 0, 0), dt.timedelta(days=1, seconds=3620), Decimal("1000.00")],
+            [2, "B", dt.datetime(2021, 1, 2, 0, 0, 0), dt.timedelta(days=2, seconds=7240), Decimal("2000.00")],
+            [3, "C", dt.datetime(2021, 1, 3, 0, 0, 0), dt.timedelta(days=3, seconds=7200), Decimal("3000.00")],
+            [4, "D", dt.datetime(2021, 1, 4, 0, 0, 0), dt.timedelta(days=4, seconds=10800), Decimal("4000.00")],
+            [5, "E", dt.datetime(2021, 1, 5, 0, 0, 0), dt.timedelta(days=5, seconds=14400), Decimal("5000.00")]
         ]
     }
 
@@ -27,12 +28,12 @@ def test_write_qvd_file(tmp_path):
     assert df is not None
     assert df.shape is not None
     assert df.shape[0] == 5
-    assert df.shape[1] == 4
+    assert df.shape[1] == 5
     assert df.columns is not None
-    assert len(df.columns) == 4
+    assert len(df.columns) == 5
     assert df.data is not None
     assert len(df.data) == 5
-    assert df.head(2).shape == (2, 4)
+    assert df.head(2).shape == (2, 5)
 
     df.to_qvd(str(tmp_path / "written.qvd"))
 
@@ -44,17 +45,22 @@ def test_write_qvd_file(tmp_path):
     assert written_df is not None
     assert written_df.shape is not None
     assert written_df.shape[0] == 5
-    assert written_df.shape[1] == 4
+    assert written_df.shape[1] == 5
 
     written_columns = written_df.to_dict()["columns"]
     written_data = written_df.to_dict()["data"]
 
-    assert written_columns == ["Key", "Value", "Timestamp", "Duration"]
-    assert written_data[0] == [1, "A", dt.datetime(2021, 1, 1, 0, 0, 0), dt.timedelta(days=1, seconds=3620)]
-    assert written_data[1] == [2, "B", dt.datetime(2021, 1, 2, 0, 0, 0), dt.timedelta(days=2, seconds=7240)]
-    assert written_data[2] == [3, "C", dt.datetime(2021, 1, 3, 0, 0, 0), dt.timedelta(days=3, seconds=7200)]
-    assert written_data[3] == [4, "D", dt.datetime(2021, 1, 4, 0, 0, 0), dt.timedelta(days=4, seconds=10800)]
-    assert written_data[4] == [5, "E", dt.datetime(2021, 1, 5, 0, 0, 0), dt.timedelta(days=5, seconds=14400)]
+    assert written_columns == ["Key", "Value", "Timestamp", "Duration", "Salary"]
+    assert written_data[0] == [1, "A", dt.datetime(2021, 1, 1, 0, 0, 0),
+                               dt.timedelta(days=1, seconds=3620), Decimal("1000.00")]
+    assert written_data[1] == [2, "B", dt.datetime(2021, 1, 2, 0, 0, 0),
+                               dt.timedelta(days=2, seconds=7240), Decimal("2000.00")]
+    assert written_data[2] == [3, "C", dt.datetime(2021, 1, 3, 0, 0, 0),
+                               dt.timedelta(days=3, seconds=7200), Decimal("3000.00")]
+    assert written_data[3] == [4, "D", dt.datetime(2021, 1, 4, 0, 0, 0),
+                               dt.timedelta(days=4, seconds=10800), Decimal("4000.00")]
+    assert written_data[4] == [5, "E", dt.datetime(2021, 1, 5, 0, 0, 0),
+                               dt.timedelta(days=5, seconds=14400), Decimal("5000.00")]
 
 def test_write_qvd_file_as_binary_file_stream(tmp_path):
     """